1d

+++ backend/src/main/java/com/backend/comment/service/impl/CommentServiceImpl.java
@@ -4,6 +4,7 @@ import com.backend.comment.domain.Comment;
 import com.backend.comment.domain.FindRequestHostDetailDto;
 import com.backend.comment.mapper.CommentMapper;
 import com.backend.comment.service.CommentService;
+import com.backend.file.domain.File;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.Authentication;
@@ -20,6 +21,7 @@ import java.io.IOException;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 
 @Service
 @Transactional(rollbackFor = Exception.class)
@@ -98,32 +100,23 @@ public class CommentServiceImpl implements CommentService {
         // 코멘트 파일 첨부
         if (files != null) {
             for (MultipartFile file : files) {
-                // db에 파일 저장
-                commentMapper.insertFileList(comment.getCommentId(), file.getOriginalFilename());
-                // 실제 파일 저장
-//                String dir = STR."C:/Temp/prj3p/\{comment.getCommentId()}"; // 부모 디렉토리(폴더)
-//                File dirFile = new File(dir);
-//                if (!dirFile.exists()) {
-//                    dirFile.mkdirs();
-//                }
-//
-//
-//                // 파일 경로
-//                String path = STR."C:/Temp/prj3p/\{comment.getCommentId()}/\{file.getOriginalFilename()}";
-//                // 저장 위치 명시
-//                File destination = new File(path);
-//                // transferTo : 인풋스트림, 아웃풋스트림을 꺼내서 하드디스크에 저장
-//                file.transferTo(destination); // checked exception 처리
+                if (!file.isEmpty()) {
 
-                // 실제 파일 저장(s3)
-                String key = STR."prj3/\{comment.getDivision()}/\{comment.getCommentId()}/\{file.getOriginalFilename()}";
-                PutObjectRequest objectRequest = PutObjectRequest.builder()
-                        .bucket(bucketName)
-                        .key(key)
-                        .acl(ObjectCannedACL.PUBLIC_READ)
-                        .build();
+                    // 실제 파일 저장(s3)
+                    String fileName = file.getOriginalFilename();
+                    String fullPath = "prj3/" + "REVIEW" + "/" + comment.getCommentId() + "/" + fileName;
 
-                s3Client.putObject(objectRequest, RequestBody.fromInputStream(file.getInputStream(), file.getSize()));
+                    PutObjectRequest objectRequest = PutObjectRequest.builder()
+                            .bucket(bucketName)
+                            .key(fullPath)
+                            .acl(ObjectCannedACL.PUBLIC_READ)
+                            .build();
+                    s3Client.putObject(objectRequest,
+                            RequestBody.fromInputStream(file.getInputStream(), file.getSize()));
+
+                    // db에 파일 저장
+                    commentMapper.insertFileList(comment.getCommentId(), fullPath);
+                }
             }
         }
     }
@@ -160,32 +153,27 @@ public class CommentServiceImpl implements CommentService {
         // List<Comment> comments = commentMapper.selectAllBySpaceId(spaceId);
         List<Comment> comments = commentMapper.selectAllBySpaceIdForReview(spaceId, offset);
         for (Comment comment : comments) {
-            // fileNames에서 파일 이름 조회
-            List<String> fileNames = commentMapper.selectByFileNameByCommentId(comment.getCommentId());
-            // 파일 경로 저장
-//            List<com.backend.file.domain.File> files = fileNames.stream()
-//                    .map(fileName -> {
-//                        var fl = new com.backend.file.domain.File();
-//                        fl.setFileName(fileName);
-//                        fl.setSrc(STR." http://172.27.128.1:8888/\{comment.getCommentId()}/\{fileName}");
-//                        return fl;
-//                    })
-//                    .toList();
+            // s3에서 파일 이름 조회
+            // List<String> files = commentMapper.selectByFileNameByCommentId(comment.getCommentId());
+            List<File> files = commentMapper.selectByFileNameByCommentIdForS3(comment.getCommentId());
+            if (files != null && !files.isEmpty()) {
+                List<File> filesWithUrls = files.stream()
+                        .map(file -> {
+                            var fl = new File();
+                            String fullPath = file.getFileName();
+                            String fileUrl = s3Client.utilities().getUrl(builder ->
+                                    builder.bucket(bucketName).key(file.getFileName())).toExternalForm();
+                            fl.setSrc(fileUrl);
+                            fl.setFileName(fullPath);
+                            System.out.println("list의 fileUrl = " + fileUrl);
+                            System.out.println("list의 fullPath = " + fullPath);
+                            return fl;
+                        }).collect(Collectors.toList());
+                // 댓글에 첨부 파일 목록 저장
+                comment.setCommentFilesLists(filesWithUrls);
 
-            // s3에서 파일 조회
-            List<com.backend.file.domain.File> files1 = fileNames.stream()
-                    .map(fileName2 ->
-                    {
-                        var fl2 = new com.backend.file.domain.File();
-                        fl2.setFileName(fileName2);
-                        fl2.setSrc(STR."\{srcPrefix}\{comment.getCommentId()}/\{fileName2}");
-                        return fl2;
-                    })
-                    .toList();
-            comment.setCommentFilesLists(files1);
-
-            // 댓글에 첨부 파일 목록 저장
-            // comment.setCommentFilesLists(files);
+                System.out.println("list의 filesWithUrls = " + filesWithUrls);
+            }
         }
         System.out.println("comments = " + comments);
 
@@ -194,34 +182,26 @@ public class CommentServiceImpl implements CommentService {
         result.put("pageInfo", pageInfo);
 
         return result;
-        // return comments;
-        // commentMapper.selectAllBySpaceId(spaceId);
     }
 
     @Override
     public void deleteReview(Comment comment) {
         // file명 조회
-        List<String> fileNames = commentMapper.selectByFileNameByCommentId(comment.getCommentId());
-        // disk에 있는 file 삭제
-//        String dir = STR."C:/Temp/prj3p/\{comment.getCommentId()}";
-//        for (String fileName : fileNames) {
-//            File file = new File(dir + fileName);
-//            file.delete();
-//        }
-//        // 필요없는 부모 디렉토리 삭제
-//        File dirFile = new File(dir);
-//        if (dirFile.exists()) {
-//            dirFile.delete();
-//        }
+        // List<String> files = commentMapper.selectByFileNameByCommentId(comment.getCommentId());
+        List<File> files = commentMapper.selectByFileNameByCommentIdForS3(comment.getCommentId());
 
-        // s3에 있는 file
-        for (String fileName : fileNames) {
-            String key = STR."prj3/\{comment.getDivision()}/\{comment.getCommentId()}/\{fileName}";
-            DeleteObjectRequest objectRequest = DeleteObjectRequest.builder()
-                    .bucket(bucketName)
-                    .key(key)
-                    .build();
-            s3Client.deleteObject(objectRequest);
+        // s3에 있는 file 삭제
+        if (files != null) {
+            for (File file : files) {
+                String fullPath = file.getFileName();
+                System.out.println("delete의 fullPath = " + fullPath);
+
+                DeleteObjectRequest objectRequest = DeleteObjectRequest.builder()
+                        .bucket(bucketName)
+                        .key(fullPath)
+                        .build();
+                s3Client.deleteObject(objectRequest);
+            }
         }
 
         // file 테이블 지움
@@ -235,20 +215,19 @@ public class CommentServiceImpl implements CommentService {
         // 첨부된 파일 삭제
         if (removeFileList != null && removeFileList.size() > 0) {
             for (String fileName : removeFileList) {
-//                String path = STR."C:/Temp/prj3p/\{comment.getCommentId()}/\{fileName}";
-//                File file = new File(path);
-//                file.delete();
-
                 // s3의 파일 삭제
-                String key = STR."prj3/\{comment.getDivision()}/\{comment.getCommentId()}/\{fileName}";
-                DeleteObjectRequest deleteObjectRequest = DeleteObjectRequest.builder()
+                String fullPath = fileName;
+                System.out.println("update의 delete의 fileName = " + fileName);
+                System.out.println("update의 delete의 fullPath = " + fullPath);
+
+                DeleteObjectRequest objectRequest = DeleteObjectRequest.builder()
                         .bucket(bucketName)
-                        .key(key)
+                        .key(fullPath)
                         .build();
-                s3Client.deleteObject(deleteObjectRequest);
+                s3Client.deleteObject(objectRequest);
 
                 // db records 삭제
-                commentMapper.deleteByCommentIdAndName(comment.getCommentId(), fileName);
+                commentMapper.deleteByCommentIdAndName(comment.getCommentId(), fullPath);
             }
         }
 
@@ -259,29 +238,21 @@ public class CommentServiceImpl implements CommentService {
             // 탐색
             for (MultipartFile file : addFileList) {
                 // 파일명을 얻어서 덮어씀
-                String fileName = file.getOriginalFilename();
+                String fileName = "prj3/" + "REVIEW" + "/" + comment.getCommentId() + "/" + file.getOriginalFilename();
                 if (!fileNameList.contains(fileName)) {
                     // 중복되지 않은 파일만 db에 추가
-                    commentMapper.insertFileList(comment.getCommentId(), file.getOriginalFilename());
+                    commentMapper.insertFileList(comment.getCommentId(), fileName);
                 }
-                // disk에 쓰기
-                // 파일이 원래 없는 경우 부모 경로 생성
-//                File dir = new File(STR."C:/Temp/prj3p/\{comment.getCommentId()}");
-//                if (!dir.exists()) {
-//                    // 디렉토리 생성
-//                    dir.mkdirs();
-//                }
-//                String path = STR."C:/Temp/prj3p/\{comment.getCommentId()}/\{fileName}";
-//                File destination = new File(path);
-//                file.transferTo(destination);
+                System.out.println("update의 add의 fileName = " + fileName);
 
                 // s3에 쓰기(덮어쓰기)
-                String key = STR."prj3/\{comment.getDivision()}/\{comment.getCommentId()}/\{fileName}";
+                String fullPath = "prj3/" + "REVIEW" + "/" + comment.getCommentId() + "/" + file.getOriginalFilename();
                 PutObjectRequest objectRequest = PutObjectRequest.builder()
                         .bucket(bucketName)
-                        .key(key)
+                        .key(fullPath)
                         .acl(ObjectCannedACL.PUBLIC_READ)
                         .build();
+
                 s3Client.putObject(objectRequest, RequestBody.fromInputStream(file.getInputStream(), file.getSize()));
             }
         }
